Separate Users and Roles entities
