+++ express-api/src/typeorm/Entities/AccessRequests.ts
@@ -1,5 +1,6 @@
 import { BaseEntity } from '@/typeorm/Entities/abstractEntities/BaseEntity';
-import { Users, Roles } from '@/typeorm/Entities/Users_Roles_Claims';
+import { Users } from '@/typeorm/Entities/Users';
+import { Roles } from '@/typeorm/Entities/Roles';
 import { Entity, Column, PrimaryGeneratedColumn, ManyToOne, Index, JoinColumn } from 'typeorm';
 import { Agencies } from './Agencies';
