+++ express-api/src/typeorm/Entities/abstractEntities/BaseEntity.ts
@@ -1,19 +1,26 @@
-import type { Users } from '@/typeorm/Entities/Users_Roles_Claims';
+import type { Users } from '@/typeorm/Entities/Users';
+import { UUID } from 'crypto';
 import { Column, CreateDateColumn, ManyToOne, JoinColumn, Index, Relation } from 'typeorm';
 
 export abstract class BaseEntity {
+  @Column({ name: 'CreatedById' })
+  CreatedById: UUID;
+
   @ManyToOne('Users', 'Users.Id')
   @JoinColumn({ name: 'CreatedById' })
   @Index()
-  CreatedById: Relation<Users>;
+  CreatedBy: Relation<Users>;
 
   @CreateDateColumn()
   CreatedOn: Date;
 
+  @Column({ name: 'UpdatedById', nullable: true })
+  UpdatedById: UUID;
+
   @ManyToOne('Users', 'Users.Id', { nullable: true })
   @JoinColumn({ name: 'UpdatedById' })
   @Index()
-  UpdatedById: Relation<Users>;
+  UpdatedBy: Relation<Users>;
 
   @Column({ type: 'timestamp', nullable: true })
   UpdatedOn: Date;
