+++ express-api/tests/testUtils/factories.ts
@@ -1,10 +1,11 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { AccessRequests } from '@/typeorm/Entities/AccessRequests';
 import { Agencies } from '@/typeorm/Entities/Agencies';
-import { Users, Roles as RolesEntity } from '@/typeorm/Entities/Users_Roles_Claims';
+import { Users } from '@/typeorm/Entities/Users';
 import { faker } from '@faker-js/faker';
 import { UUID } from 'crypto';
 import { Request, Response } from 'express';
+import { Roles as RolesEntity } from '@/typeorm/Entities/Roles';
 
 export class MockRes {
   statusValue: any;
@@ -78,7 +79,9 @@ export const produceUser = (): Users => {
     CreatedOn: faker.date.anytime(),
     UpdatedOn: faker.date.anytime(),
     UpdatedById: undefined,
+    UpdatedBy: undefined,
     CreatedById: undefined,
+    CreatedBy: undefined,
     Id: id,
     DisplayName: faker.company.name(),
     FirstName: faker.person.firstName(),
@@ -93,9 +96,11 @@ export const produceUser = (): Users => {
     Note: '',
     LastLogin: faker.date.anytime(),
     ApprovedById: undefined,
+    ApprovedBy: undefined,
     ApprovedOn: undefined,
     KeycloakUserId: faker.string.uuid() as UUID,
-    UserRoles: [],
+    Role: produceRole(),
+    RoleId: undefined,
     Agency: produceAgency(id),
     AgencyId: undefined,
   };
@@ -110,28 +115,33 @@ export const produceRequest = (): AccessRequests => {
     RoleId: undefined,
     AgencyId: produceAgency(),
     CreatedById: undefined,
+    CreatedBy: undefined,
     CreatedOn: faker.date.anytime(),
     UpdatedById: undefined,
+    UpdatedBy: undefined,
     UpdatedOn: faker.date.anytime(),
   };
   return request;
 };
 
-export const produceAgency = (userId?: string): Agencies => {
+export const produceAgency = (code?: string): Agencies => {
   const agency: Agencies = {
-    Id: userId ?? faker.string.numeric(6),
+    Id: faker.number.int(),
     Name: faker.company.name(),
     IsDisabled: false,
     SortOrder: 0,
     Description: '',
     ParentId: undefined,
+    Code: code ?? faker.string.alpha({ length: 4 }),
     Email: faker.internet.email(),
     SendEmail: false,
     AddressTo: '',
     CCEmail: faker.internet.email(),
     CreatedById: undefined,
     CreatedOn: new Date(),
+    CreatedBy: undefined,
     UpdatedById: undefined,
+    UpdatedBy: undefined,
     UpdatedOn: new Date(),
     Users: [],
   };
@@ -143,7 +153,9 @@ export const produceRole = (): RolesEntity => {
     CreatedOn: faker.date.anytime(),
     UpdatedOn: faker.date.anytime(),
     UpdatedById: undefined,
+    UpdatedBy: undefined,
     CreatedById: undefined,
+    CreatedBy: undefined,
     Id: faker.string.uuid() as UUID,
     Name: faker.company.name(),
     IsDisabled: false,
@@ -151,7 +163,6 @@ export const produceRole = (): RolesEntity => {
     SortOrder: 0,
     KeycloakGroupId: faker.string.uuid() as UUID,
     IsPublic: false,
-    UserRoles: [],
-    RoleClaims: [],
+    Users: [],
   };
 };
